(PC-42034) refactor(offer): artist section before accessibility section
