+++ src/features/offer/components/OfferBody/OfferBody.tsx
@@ -9,6 +9,7 @@ import { useAuthContext } from 'features/auth/context/AuthContext'
 import { isCurrentBeneficiary } from 'features/auth/helpers/checkStatusType'
 import { UseNavigationType, UseRouteType } from 'features/navigation/navigators/RootNavigator/types'
 import { OfferAbout } from 'features/offer/components/OfferAbout/OfferAbout'
+import { OfferAccessibility } from 'features/offer/components/OfferAccessibility/OfferAccessibility'
 import { OfferArtists } from 'features/offer/components/OfferArtists/OfferArtists'
 import { OfferArtistsSection } from 'features/offer/components/OfferArtistsSection/OfferArtistsSection'
 import { ProposedBySection } from 'features/offer/components/OfferBody/ProposedBySection/ProposedBySection'
@@ -247,6 +248,12 @@ export const OfferBody: FunctionComponent<Props> = ({
             offerSubcategoryId={offer.subcategoryId}
             onPlaylistItemPress={handleOnArtistPlaylistItemPress}
           />
+        </MarginContainer>
+      ) : null}
+
+      {shouldDisplayAccessibilitySection ? (
+        <MarginContainer gap={0}>
+          <OfferAccessibility accessibility={offer.accessibility} />
         </MarginContainer>
       ) : null}
